Implement SchedulingStrategy and derivative classes so they work with the Scheduler refactor
